Add option to retrieve issues from GitHub
Added the option -t to instruct the bot to retrieve information about
an issue. Requires a GitHub API token.

The rejoinfile is no longer truncated and rewritten, but is replaced
by a previously written temporary file. It now also contains the
channel that was passed on the command line, because the rejoinfile is
rewritten on each chanjoin event, rather than on calls to
join_channel().

Updated the manual.