Handle nick-in-use automatically.
Do not call a nickname_in_use_error() method when our nick name is
already in use. Instead, use the
POE::Component::IRC::Plugin::NickReclaim plugin to handle it and also
reclaim the nick when it becaomes available again.

The nick() method now tries to change our nick on the server, if we
are already connected, rather than just set the nick for the next time
we connect.