Various fixes and improvements
Set version number to 0.3.

When writing the mapfile to store current session data, write
information about all channels, not just those that have associated
repositories. (Because there is other data to remember for a channel,
such as the delay and whether expansion is on or off.)

In add_repositories(), do not rewrite the mapfile unless new
repositories were actually added.

Remove a duplicate log message.

When a reference to an issue cannot be expanded, because no matching
repository is known, and the issue was part of a message explicitly
addressed to us, then reply that the issue couldn't be expanded,
rather then stay quiet.

When replying to a status request with a list of known repositories,
put the word "and" instead of a comma before the last repo in the
list.

When adding nicks to ignore, reply with "OK" only when there were no
warnings about already ignored nicks.

Simplified some regexps in said(), because leading and trailing white
space is already stripped before the method is called.

Expand the help text with help on each individual command.