Try to make some error message more helpful
Improve some error messages when GitHub responds with 404: It may
indicate a lack of access rights rather than a nonexistent
repository.

Print an error if an access token is needed for an operation but
wasn't provided.

Print an error if the bot was addressed explicitly, but with text that
was neither a command nor contained any references to expand.

Fixed a bug where a possibly uninitialized variable was checked to
determine if an alias was already set.