Nicer output.
Accept (but ignore) an empty repo name in the mapfile. (Not sure where
the empty names comes from. Maybe the result of an earlier bug?)

Only write information to the mapfile for channels that have
information: nonempty list of repos, a non-default delay, etc.

Sort the mapfile for easier debugging.

When showing a closed issue, prefix it with ‘CLOSE’ instead of putting
‘[closed]’ after the number, because the latter may be confused with
actual text from the issue title, such as ‘[css-fonts]’.

When showing an issue, put ‘by’ before the name of the creator. Also
put each label in ‘[]’, to better show it is a label and not part of
the title.