Copied some changes from the main branch:
Don't expand a name ("@foo") if it is preceded by another "@" ("@@foo").

Log also when an issue number was expanded to a URL without being looked up.

Test for /^@/ instead of /@/ (for a very small gain in speed).

Allow $self->{suspend_names} to be 0 as well as undef or
non-existent. (Cannot currently occur, but maybe in the future.)