Close and reopen now also work for discussions.
This required replacing some more calls to the REST API with calls to
the GraphQL API.

URLs to pull requests and to discussions are now also recognized and
cause gb to look up a summary, just like for issues.

Added the get_github_id_and_type() function, because several other
functions need to look up the node ID of issues. (Unlike in the REST
API, in the GraphQL API the repo name and issue number are not enough
to change an issue. You need to look up the node ID and use that to
change the issue.)